[persistence] separate the cmdlogfile module from cmdlogbuf.c
