+++ engines/default/checkpoint.c
@@ -28,6 +28,7 @@
 #include "checkpoint.h"
 #include "mc_snapshot.h"
 #include "cmdlogbuf.h"
+#include "cmdlogfile.h"
 
 #define CHKPT_MAX_FILENAME_LENGTH  255
 #define CHKPT_FILE_NAME_FORMAT     "%s/%s%"PRId64
@@ -265,7 +266,7 @@ static bool do_checkpoint_needed(chkpt_st *cs)
 {
     struct engine_config *config = cs->config;
     size_t snapshot_file_size = cs->lastsize;
-    size_t cmdlog_file_size   = cmdlog_file_getsize();
+    size_t cmdlog_file_size   = cmdlog_get_file_size();
     size_t min_logsize        = config->chkpt_interval_min_logsize;
     int    pct_snapshot       = config->chkpt_interval_pct_snapshot;
