+++ engines/default/cmdlogbuf.h
@@ -21,22 +21,19 @@
 #include "cmdlogmgr.h"
 #include "cmdlogrec.h"
 
-/* external log functions */
+/* global log flush mutex */
+extern pthread_mutex_t log_flush_lock;
+extern pthread_cond_t  log_flush_cond;
+
+/* external log buffer functions */
 void cmdlog_buff_write(LogRec *logrec, log_waiter_t *waiter, bool dual_write);
 void cmdlog_buff_flush(LogSN *upto_lsn);
-void cmdlog_file_sync(void);
 
 /* FIXME: remove later, if not used */
 //void log_get_write_lsn(LogSN *lsn);
-void cmdlog_get_flush_lsn(LogSN *lsn);
-void cmdlog_get_fsync_lsn(LogSN *lsn);
+void   cmdlog_get_flush_lsn(LogSN *lsn);
+size_t cmdlog_get_file_size(void);
 
-int               cmdlog_file_open(char *path);
-size_t            cmdlog_file_getsize(void);
-void              cmdlog_file_close(bool chkpt_success);
-void              cmdlog_file_init(void);
-void              cmdlog_file_final(void);
-int               cmdlog_file_apply(void);
 void              cmdlog_complete_dual_write(bool success);
 ENGINE_ERROR_CODE cmdlog_buf_init(struct default_engine *engine);
 void              cmdlog_buf_final(void);
